cf/logging_test/requirements.txt
google-cloud-logging==2.3.1